+++ src/test/java/services/CommandServiceImplTest.java
@@ -16,6 +16,8 @@ import services.report.DailyReportBuilder;
 import services.shortage.BeverageQuantityChecker;
 import services.shortage.EmailNotifier;
 
+import java.math.BigDecimal;
+
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
@@ -46,7 +48,7 @@ class CommandServiceImplTest {
     @DisplayName("should translate a coffee order with sugar and sufficient money")
     void shouldOrderCoffeeWithSugar() {
         final Order order = new Order(OrderType.COFFEE, 2, false);
-        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(1.0);
+        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(BigDecimal.valueOf(1.0));
         commandService.sendOrderToDrinkMaker(order);
         final String expectedOrder = "C:2:0";
 
@@ -60,7 +62,7 @@ class CommandServiceImplTest {
     @DisplayName("should translate a hot coffee order with sugar and sufficient money")
     void shouldOrderHotCoffeeWithSugar() {
         final Order order = new Order(OrderType.COFFEE, 2, true);
-        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(1.0);
+        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(BigDecimal.valueOf(1.0));
         commandService.sendOrderToDrinkMaker(order);
         final String expectedOrder = "Ch:2:0";
 
@@ -74,7 +76,7 @@ class CommandServiceImplTest {
     @DisplayName("should translate a orange juice order with sufficient money")
     void shouldOrderOrangeJuice() {
         final Order order = new Order(OrderType.ORANGE_JUICE, 0, false);
-        when(moneyChecker.isAffordable(OrderType.ORANGE_JUICE)).thenReturn(1.0);
+        when(moneyChecker.isAffordable(OrderType.ORANGE_JUICE)).thenReturn(BigDecimal.valueOf(1.0));
         commandService.sendOrderToDrinkMaker(order);
         final String expectedOrder = "O::";
 
@@ -88,7 +90,7 @@ class CommandServiceImplTest {
     @DisplayName("should translate a coffee order with sugar and send message")
     void shouldNotOrderCoffeeWithoutMoneyAndSendMsg() {
         final Order order = new Order(OrderType.COFFEE, 2, false);
-        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(-0.5);
+        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(BigDecimal.valueOf(-0.5));
         commandService.sendOrderToDrinkMaker(order);
 
         verify(drinkMaker).sendMessage("M:Monney is missing: 0.5 €");
@@ -99,7 +101,7 @@ class CommandServiceImplTest {
     @Test
     @DisplayName("should translate a tee order without sugar and sufficient money")
     void shouldTranslateTeaOrderWithoutSugarToString() {
-        when(moneyChecker.isAffordable(OrderType.TEA)).thenReturn(0.0);
+        when(moneyChecker.isAffordable(OrderType.TEA)).thenReturn(BigDecimal.valueOf(0));
         final Order order = new Order(OrderType.TEA, 0, false);
         final String expectedOrder = "T::";
         commandService.sendOrderToDrinkMaker(order);
@@ -124,9 +126,9 @@ class CommandServiceImplTest {
     @Test
     @DisplayName("should add money to coffee machine")
     void shouldAddMoneyToMachine() {
-        commandService.insertMoney(5.5);
+        commandService.insertMoney(BigDecimal.valueOf(5.5));
 
-        verify(moneyChecker).insertMoney(5.5);
+        verify(moneyChecker).insertMoney(BigDecimal.valueOf(5.5));
         verifyNoMoreInteractions(moneyChecker);
     }
 
@@ -144,7 +146,7 @@ class CommandServiceImplTest {
     @Test
     @DisplayName("should not make a drink with shortage and send message")
     void shouldNotMakeDrinkWithShortageAndSendMessage() {
-        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(0.0);
+        when(moneyChecker.isAffordable(OrderType.COFFEE)).thenReturn(BigDecimal.valueOf(0));
         when(beverageQuantityChecker.isEmpty("C::")).thenReturn(true);
         final Order order = new Order(OrderType.COFFEE, 0, false);
